Use Markdown README for PyPI
The new version of PyPI and Warehouse added support for PEP 566,
which adds the option "long_description_content_type" to specify the
format of the long description used on the PyPI page. This removes
the need to convert it to RST using pypandoc, which (for this package
at least) was not formatting correctly.